+++ src/main/kotlin/com/d1t/dastargram/global/config/entity/JpaAuditingConfiguration.kt
@@ -1,4 +1,4 @@
-package com.d1t.dastargram.global.config
+package com.d1t.dastargram.global.config.entity
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
